+++ packages/firebase_tools_common/lib/firebase_project.dart
@@ -7,6 +7,7 @@ export 'src/firebase_project.dart'
         FirebaseProjectOptions,
         FirebaseProjectBuilder,
         FirebaseProjectActionController,
+        firebaseDeployCommand,
         firebaseFunctionsDeployOnly,
         firebaseFolderProjectId,
         firebaseRcContentProjectId;